requirements.txt
pyrtlsdr
numpy
matplotlib
scipy
flask
scikit-learn
tensorflow>=2.12.0,<2.16.0
